last changes

+++ core-service/src/main/java/ru/vasili_zlobin/web_market/core/model/Product.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "products")
 @Data
@@ -18,9 +20,9 @@ public class Product {
     private String title;
 
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
 
-    public Product(String title, Double price) {
+    public Product(String title, BigDecimal price) {
         this.title = title;
         this.price = price;
     }
